rsoi-2022-lab3-fault-tolerance-malshevskikh-master/rsoi-2022-lab3-fault-tolerance-malshevskikh-master/source_v1/FlightService/requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.2.1
Babel==2.8.0
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.1.1
circuitbreaker==1.4.0
click==8.0.3
cloudpickle==2.0.0
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
dask==2021.10.0
DateTime==4.7
debugpy==1.6.4
decorator==4.4.2
defusedxml==0.7.1
distributed==2021.10.0
Django==4.1.1
djangorestframework==3.14.0
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
fonttools==4.38.0
fsspec==2021.10.1
HeapDict==1.0.1
httpie==3.2.1
idna==2.10
ipykernel==6.19.2
ipython==8.7.0
jedi==0.18.2
Jinja2==3.0.2
jupyter_client==7.4.8
jupyter_core==5.1.0
kiwisolver==1.4.4
locket==0.2.1
MarkupSafe==2.0.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
msgpack==1.0.2
multidict==6.0.2
nest-asyncio==1.5.6
numpy==1.19.4
packaging==21.0
parso==0.8.3
partd==1.2.0
pexpect==4.8.0
pi==0.1.2
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.11
platformdirs==2.6.0
prompt-toolkit==3.0.36
psutil==5.8.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.1

PySocks==1.7.1
python-dateutil==2.8.2
pytz==2020.1
PyYAML==6.0
pyzmq==24.0.1
qtwidgets==0.17
requests==2.24.0
requests-toolbelt==0.9.1
rich==12.6.0

schedule==1.1.0
shiboken2==5.15.1
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.3
stack-data==0.6.2
tblib==1.7.0
tkcalendar==1.6.1
toolz==0.11.1
tornado==6.2
traitlets==5.7.1
typing_extensions==4.3.0
urllib3==1.25.10
views==0.3
wcwidth==0.2.5
yarg==0.1.9
zict==2.0.0
zope.interface==5.5.2
